fixup! Add webkitgtk_minibrowser runs to taskcluster daily. (#19466)
